% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alrc_as_made.R
\name{alrc_as_made}
\alias{alrc_as_made}
\title{Filter data from the Australian Law Reform Commission's "As Made Commonwealth Legislation" dataset}
\usage{
alrc_as_made(prin_amend = NULL, leg_type = NULL)
}
\arguments{
\item{prin_amend}{A character string indicating whether to filter by "Principal" or "Amending" acts.
Default is NULL, which does not apply this filter.}

\item{leg_type}{A character string indicating the legislation type to filter by.
Options are "Acts", "Regulations", or "Other LI". Default is NULL, which does not
apply this filter.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A filtered data object.
}\if{html}{\out{</div>}}
}
\description{
This function filters data from the Australian Law Reform Commission's "As Made"
dataset based on the specified principal/amending and legislation type. The default
dataset is the full "As Made Commonwealth Legislation" dataset.
}
\examples{
alrc_as_made(prin_amend = "Principal", leg_type = "Acts")
alrc_as_made(leg_type = "Other LI")
}
