% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alrc_rels.R
\name{alrc_relationship_data}
\alias{alrc_relationship_data}
\title{Provide linkage data from the Australian Law Reform Commission's "Legislative relationships" datasets}
\usage{
alrc_relationship_data(rel_type = "amendments")
}
\arguments{
\item{rel_type}{A character string indicating the type of relationship data to filter.
Options are "amendments", "repeals", "enabled_by", or "modifications".
Default is "amendments".}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    A filtered data object.
}\if{html}{\out{</div>}}
}
\description{
This function provides data from the Australian Law Reform Commission's
"Legislative relationships" datasets based on the specified relationship type.
The default dataset is the amending acts legislative relationships data.
}
\details{
Linkages from one piece of legislation to the next may be made using the \code{fromId}
and \code{toId} columns (from alrc_relationship_data) and the \code{id} column (from
alrc_as_made)
}
\examples{
alrc_relationship_data(rel_type = "repeals")
alrc_relationship_data()
}
