% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAmendingActs.R
\name{getAmendingActs}
\alias{getAmendingActs}
\title{get principal acts and details}
\usage{
getAmendingActs()
}
\value{
a tibble
}
\description{
get principal acts and details
}
\examples{
getAmendingActs()
}
