% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrincipalActs.R
\name{getPrincipalActs}
\alias{getPrincipalActs}
\title{get principal acts and details}
\usage{
getPrincipalActs()
}
\value{
a tibble
}
\description{
get principal acts and details
}
\examples{
getPrincipalActs()
}
